% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuiper.2samp.R
\name{kuiper.2samp}
\alias{kuiper.2samp}
\title{2-sample Kuiper Test Function: performs Kuiper Test for two sets samples of observations}
\usage{
kuiper.2samp(x, y)
}
\arguments{
\item{x}{an array of sample observations}

\item{y}{the other array of sample observations}
}
\value{
Kuiper test statistic and p-value
}
\description{
2-sample Kuiper Test Function: performs Kuiper Test for two sets samples of observations
}
\note{
The computation of the p-value takes references from Paltani(2004)
 which states that the functions (in the set of four formulas) never underestimates the false positive probability
 however it can be a bit high when the sample size in the range of 40 to 50
 a factor of 1.5 is quoted at the 1e-7 level
}
\examples{
kuiper.2samp(rnorm(1e3),rnorm(1e3))
}
\references{
Kuiper, N. H. (1960). "Tests concerning random points on a circle". Proceedings of the Koninklijke Nederlandse Akademie van Wetenschappen, Series A. 63: 38-47.
 Paltani, S., "Searching for periods in X-ray observations using Kuiper's test. Application to the ROSAT PSPC archive", Astronomy and Astrophysics, v.240, p.789-790, 2004.
}
